requirements.txt
amqp==5.3.1
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.8.1
billiard==4.2.1
cachetools==4.2.2
celery==5.5.3
certifi==2025.6.15
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==13.15
pytz==2025.2
redis==6.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.13
